requirements.txt
# requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
jinja2==3.1.4
sqlalchemy==2.0.35
pydantic==2.9.2
email-validator==2.2.0
python-multipart==0.0.9
Pillow==11.0.0
pytesseract==0.3.10